[API CHANGE] MathTools only used in a single package in structgen, move it there and make non-public.
