+++ tool/structgen/src/test/java/org/openscience/cdk/structgen/MathToolsTest.java
@@ -17,11 +17,10 @@
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
  *
  */
-package org.openscience.cdk.math;
+package org.openscience.cdk.structgen;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.openscience.cdk.math.MathTools;
 import org.openscience.cdk.CDKTestCase;
 
 /**
